Split TrialArtifactCards component
